0.04 [#0]
→  Adjusted the opacity of the background of the Simple theme
→  Fixed the inventory closing and opening on the same frame if Reduced Motion is enabled
→  Fixed the door in Debug World being invisible
→  Fixed the collision of the Statue of Liberty in Macacolandia being larger than expected
→  Fixed the collision of the horn in Debug World being larger than expected
→  Fixed the title on the main menu’s language selection level not updating to the highlighted option
→  Fixed a crash when selecting the theme that’s already equipped
→  Fixed the depth of the player not updating after changing rooms
and more. I fogot

+++ scripts/scr_config/scr_config.gml
@@ -96,8 +96,8 @@ function fn_config_setup()
 	
 	// Graphics
 	global.config_fscr = false; // Fullscreen
-	global.config_wndSc = 1; // Window Scale
-	global.config_vsync = false; // Vsync (display_reset)
+	//global.config_wndSc = 1; // Window Scale
+	//global.config_vsync = false; // Vsync (display_reset)
 	global.config_lowGfx = false; // Low Graphics
 	global.config_hideCsr = false; // Hide Cursor
 	global.config_showFps = false; // Show FPS
@@ -252,5 +252,6 @@ function fn_config_file_load()
 }
 function fn_config_file_erase()
 {
-	file_delete(global.config_file.name);
+	if (file_exists(global.config_file.name) == true)
+		file_delete(global.config_file.name);
 }
